OV-113: * rename time-axis helper
