+++ frontend/src/bundles/studio/components/timeline/timeline-view/timeline-view.tsx
@@ -2,7 +2,7 @@ import { useTimelineContext } from 'dnd-timeline';
 
 import { Box } from '~/bundles/common/components/components.js';
 import { RowNames } from '~/bundles/studio/enums/enums.js';
-import { timeAxisMarkers } from '~/bundles/studio/helpers/time-axis-markers.helper.js';
+import { timeAxisMarkers } from '~/bundles/studio/helpers/time-axis-markers.js';
 import {
     type DestinationPointer,
     type TimelineRowsWithSpan,
