+++ frontend/src/bundles/studio/helpers/time-axis-markers.ts
